@@ -1,19 +1,12 @@
-package es.alruiz.awesomeapp.ui.main;
+package es.alruiz.awesomeapp.ui.main
 
-import java.util.List;
-
-import es.alruiz.awesomeapp.objects.User;
+import es.alruiz.awesomeapp.objects.User
 
 /**
  * Created by AlfonsoRuiz on 30/10/2016.
  */
-
-public interface MainView {
-
-    void showLoading();
-
-    void hideLoading();
-
-    void setRepositoryUsers(List<User> body);
-
+interface MainView {
+    fun showLoading()
+    fun hideLoading()
+    fun setRepositoryUsers(body: List<User>)
 }
